requirements.txt
asgiref==3.3.1
Django==3.1.6
django-crispy-forms==1.11.2
pytz==2021.1
sqlparse==0.4.1
whitenoise==6.4.0
mixer~=7.1.2
djangorestframework~=3.14.0